feat(fabricatio-tool): add ping functionality and improve error handling
- Add ping method to McpManager for checking client connectivity
- Implement custom error type (McpError) for better error handling
- Refactor mcp.rs to use new MCPManager struct
- Update dependencies and add new crates for improved functionality
