+++ Cargo.toml
@@ -36,6 +36,7 @@ members = [
     "packages/fabricatio-anki",
     "packages/fabricatio-anki/src/deck_loader",
     "packages/fabricatio-tool",
+    "packages/fabricatio-tool/crates/mcp-manager",
     "packages/fabricatio-locale",
     "packages/fabricatio-thinking",
     "packages/fabricatio-agent",
