+++ packages/fabricatio-tool/Cargo.toml
@@ -21,7 +21,7 @@ rmcp = { version = "0.3.2", features = [
 ] }
 rustpython-ast = { version = "0.4.0", features = ["visitor"] }
 rustpython-parser = { version = "0.4.0"  }
-serde = { version = "1.0.219", features = ["derive"] }
 serde_json = "1.0.142"
-tokio = { version = "1.47.1", features = ["process", "rt-multi-thread", "sync"] }
-which = "8.0.0"
+tracing-subscriber = { version = "0.3.19", features = ["env-filter"] }
+
+mcp-manager = { path = "./crates/mcp-manager" }
