+++ packages/fabricatio-tool/crates/mcp-manager/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "mcp-manager"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+futures = "0.3.31"
+rmcp = { version = "0.3.2", features = [
+    "auth",
+    "client",
+    "transport-child-process",
+    "transport-io",
+    "transport-sse-client",
+    "transport-streamable-http-client"
+] }
+serde = { version = "1.0.219", features = ["derive"] }
+serde_json = "1.0.142"
+tokio = { version = "1.47.1", features = ["process", "rt-multi-thread", "sync"] }
+which = "8.0.0"
